Move fix constants to a file
